% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix-clip-path.R
\name{fix_clip_path}
\alias{fix_clip_path}
\title{Remove \code{clip-path} element from SVG files}
\usage{
fix_clip_path(path)
}
\arguments{
\item{path}{the path of a folder of a project that may contain SVG files.}
}
\value{
A character vector of the absolute paths of the affected SVG files.
}
\description{
Some version of Firefox don't display correctly SVG files that have a
\code{clip-path} element set. This function detects and removes this element from
any SVG files included within any subfolders given as a path.
}
\details{
The diff between the files affected by this script and the original might be
significantly different, as the entire content of the file is parsed and
rewritten (using the \code{xml2} package). This has the side effect of
potentially affecting the order of the elements of the file leading to large
diffs.
}
